+++ language/tools/move-mv-llvm-compiler/src/main.rs
@@ -252,8 +252,6 @@ fn main() -> anyhow::Result<()> {
                 let disasm = module.disassemble();
                 println!("Module {} bytecode {}", modname, disasm);
             }
-            // let mod_cx =
-            //     global_cx.create_module_context(mod_id, &llmod, &entrypoint_generator, &options);
             let mod_src = module.get_source_path().to_str().expect("utf-8");
             let mod_cx = &mut global_cx.create_module_context(
                 mod_id,
